version 0.1.0-19

+++ src/getdata.c
@@ -35,6 +35,8 @@
 #include "defns.h"
 #include "extern.h"
 
+#include <stdint.h>
+
 #include "transform.h"
 #include "redefine.h"
 
@@ -283,7 +285,7 @@ DataRec GetDataRec(FILE *Df, Boolean Train)
 			{
 			    /*  Add value to list  */
 
-			    if ( MaxAttVal[Att] >= (long) AttValName[Att][0] )
+			    if ( MaxAttVal[Att] >= (long) (intptr_t) AttValName[Att][0] )
 			    {
 				XError(TOOMANYVALS, AttName[Att],
 					 (char *) AttValName[Att][0] - 1);
@@ -497,7 +499,7 @@ DataRec PredictGetDataRec(FILE *Df, Boolean Train)
 			{
 			    /*  Add value to list  */
 
-			    if ( MaxAttVal[Att] >= (long) AttValName[Att][0] )
+			    if ( MaxAttVal[Att] >= (long) (intptr_t) AttValName[Att][0] )
 			    {
 				XError(TOOMANYVALS, AttName[Att],
 					 (char *) AttValName[Att][0] - 1);
